refactor: drop the unconsumed SwiftUI wrapper layer; fix load-failure message
@SyncedQuery/@SyncedModel/SyncedResults/SyncedView had zero code callers — only
the publishers (SyncedQueryPublisher/SyncedModelPublisher) are used, by the demo
machines. Remove the speculative surface; keep + split the consumed types into
SyncLoadPhase.swift / SyncedQueryPublisher.swift / SyncedModelPublisher.swift.

Fix (red-first): the shared load driver surfaced an empty/whitespace
errorDescription verbatim and had no per-screen fallback, regressing the screens'
actionable text. Trim + empty-check like presentError, and thread a screen-supplied
fallbackMessage (projects/project/task) through the publishers.

SwiftSync 182, DemoCore 43 green.
